+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "pilar-web",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "private": true,
   "dependencies": {
     "@emotion/react": "^11.11.1",
