+++ src/pages/layouts/DashboardLayout.jsx
@@ -10,19 +10,14 @@ import {
   Avatar,
   Button,
   List,
-  ListItem,
-  ListItemButton,
   ListItemText,
-  ListItemIcon,
   Drawer,
-  Collapse,
   MenuItem as MenuItemMui,
   Popover
 } from '@mui/material';
 import MenuIcon from "@mui/icons-material/Menu"
 import { drawerMenu, popMenu } from '../../constants/menu'
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
-import ChevronRightIcon from '@mui/icons-material/ChevronRight'
+import MenuItem from '../../components/MenuItem';
 
 //6Definir las librerias arriba
 
@@ -80,70 +75,7 @@ const PopMenu = () => {
 
 const drawerWidth = 240;
 
-// 4_componente MenuItem al cual le pasamos el item en sus props
 
-const MenuItem = ({ item }) => {
-  const navigate = useNavigate()
-  const { children, title, path } = item
-  const [open, setOpen] = useState(false)
-  if (children) {
-    return (
-      <>
-        <ListItem key={item.title} onClick={() => setOpen(status => !status)} >
-          <ListItemButton>
-            <ListItemText sx={{ fontWeight: 400 }} primary={item.title}
-              disableTypography />
-          </ListItemButton>
-          {open ? (<ExpandMoreIcon />) : (<ChevronRightIcon />)}
-        </ListItem>
-        <Collapse in={open} timeout='auto' unmountOnExit>
-          <List component="div" disablePadding>
-            {
-              children.map((child, index) => {
-                return (
-                  <ListItem key={index}
-                    onClick={() => navigate(child.path)} disablePadding>
-                    <ListItemButton sx={{
-                      height: 42, fontWeight: '200',
-                      padding: th => th.spacing(0, 2.5, 0, 3)
-                    }} >
-                      <ListItemIcon sx={{
-                        width: 26, height: 26,
-                        display: 'flex', alignItems: 'center', justifyContent: 'center', padding:
-                          th => th.spacing(0, 2, 0, 0)
-                      }} >
-                        <Box
-                          component='span'
-                          sx={{
-                            width: 4, height: 4, display: 'flex',
-                            borderRadius: '50%', alignItems: 'center', justifyContent: 'center',
-                            backgroundColor: 'text.disabled'
-                          }}
-                        />
-                      </ListItemIcon>
-                      <ListItemText sx={{ fontWeight: 400 }}
-                        primary={child.title} />
-                    </ListItemButton>
-                  </ListItem>
-                )
-              })
-            }
-          </List>
-        </Collapse>
-      </>
-    )
-  }
-  return (
-    <>
-      <ListItem onClick={() => navigate(path)} >
-        <ListItemButton>
-          <ListItemText sx={{ fontWeight: 400 }} primary={title}
-            disableTypography />
-        </ListItemButton>
-      </ListItem>
-    </>
-  )
-}
 
 //3 componente Menu Definido:
 const Menu = ({ items }) => {
